+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 * Update to clojure 1.11.0-alpha2 and cheshire 5.10.1
 * Add [Nix](https://nixos.org/) and [direnv](https://direnv.net/) files for development environment
+* Add cidr, fn-and, fn-or, get-azs, select, and transform functions
 
 ## 0.1.0 (2021-05-09)